Temporarily remove some packages
Grouper features an extremely small system partition. If it is compiled without the removal of these packages there is not enough space to flash GApps. I decided to remove these packages as they should work on the data partition as well so I will add them to my GApps package which will flash them to the data partition.

Change-Id: Ic0c64362541757d21b2eebe53d42c1db447361e9